+++ js/_website/src/lib/components/Details.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-	import Lightbulb from "$lib/components/icons/Lightbulb.svelte";
 	export let summary: string;
 	export let content: string;
 
@@ -8,7 +7,7 @@
 
 <div class="container">
 	<button class="summary" class:open on:click={() => (open = !open)}
-		><span><Lightbulb /></span>{summary}</button
+		>{summary}</button
 	>
 	{#if open}
 		<div class="detail">{@html content}</div>
@@ -26,7 +25,6 @@
 		font-size: 1.125rem;
 		font-weight: 600;
 		line-height: 1.75rem;
-		color: #374151;
 		cursor: pointer;
 		display: flex;
 		align-items: center;
@@ -39,11 +37,5 @@
 		line-height: 1.5rem;
 		color: #374151;
 		text-align: justify;
-	}
-
-	span {
-		color: rgb(9, 105, 218);
-		margin-left: -3px;
-		margin-right: 0.75rem;
 	}
 </style>