+++ js/paramviewer/Index.svelte
@@ -1,19 +1,16 @@
 <script lang="ts">
+	import type { ParamViewerProps, ParamViewerEvents } from "./types";
+	import { Gradio } from "@gradio/utils";
 	import ParamViewer from "./ParamViewer.svelte";
 
-	export let value: Record<
-		string,
-		{
-			type: string;
-			description: string;
-			default: string;
-		}
-	>;
-
-	export let linkify: string[] = [];
-	export let header: string | null = null;
-	export let anchor_links = false;
-	export let max_height: number | string | undefined = undefined;
+	const props = $props();
+	const gradio = new Gradio<ParamViewerEvents, ParamViewerProps>(props);
 </script>
 
-<ParamViewer docs={value} {linkify} {header} {anchor_links} {max_height} />
+<ParamViewer
+	docs={gradio.props.value}
+	linkify={gradio.props.linkify}
+	header={gradio.props.header}
+	anchor_links={gradio.props.anchor_links}
+	max_height={gradio.props.max_height}
+/>