+++ js/video/Video.stories.svelte
@@ -46,11 +46,9 @@
 	name="Static video"
 	args={{
 		value: {
-			video: {
-				path: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
-				url: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
-				orig_name: "world.mp4"
-			}
+			path: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
+			url: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
+			orig_name: "world.mp4"
 		},
 		label: "world video",
 		show_label: true,
@@ -68,11 +66,9 @@
 	name="Static video with vertical video"
 	args={{
 		value: {
-			video: {
-				path: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world_vertical.mp4",
-				url: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world_vertical.mp4",
-				orig_name: "world_vertical.mp4"
-			}
+			path: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world_vertical.mp4",
+			url: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world_vertical.mp4",
+			orig_name: "world_vertical.mp4"
 		},
 		label: "world video",
 		show_label: true,
@@ -115,11 +111,9 @@
 		width: 400,
 		height: 400,
 		value: {
-			video: {
-				path: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
-				url: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
-				orig_name: "world.mp4"
-			}
+			path: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
+			url: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
+			orig_name: "world.mp4"
 		},
 		webcam_options: {
 			mirror: true,
@@ -132,11 +126,9 @@
 	name="Trim video"
 	args={{
 		value: {
-			video: {
-				path: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
-				url: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
-				orig_name: "world.mp4"
-			}
+			path: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
+			url: "https://gradio-static-files.s3.us-west-2.amazonaws.com/world.mp4",
+			orig_name: "world.mp4"
 		},
 		label: "world video",
 		show_label: true,