+++ js/box/CHANGELOG.md
@@ -1,4 +1,25 @@
 # @gradio/box
+## 0.2.26
+
+### Features
+
+- [#12438](https://github.com/gradio-app/gradio/pull/12438) [`25ffc03`](https://github.com/gradio-app/gradio/commit/25ffc0398f8feb43d817c02b2ab970c16de6d797) - Svelte5 migration and bugfix
+
+### Dependencies
+
+- @gradio/atoms@0.19.0
+
+## 0.2.26-dev.1
+
+### Dependency updates
+
+- @gradio/atoms@0.19.0-dev.1
+
+## 0.2.26-dev.0
+
+### Dependency updates
+
+- @gradio/atoms@0.18.2-dev.0
 
 ## 0.2.25